Unify tuning and training evaluation core
